fixed rating bar in detailsActivity

+++ app/src/main/java/edu/orangecoastcollege/cs273/wheretonext/CollegeListActivity.java
@@ -38,14 +38,16 @@ public class CollegeListActivity extends AppCompatActivity {
         mTuitionEditText = (EditText) findViewById(R.id.tuitionEditText);
         mRatingBar = (RatingBar) findViewById(R.id.collegeRatingBar);
 
-        // TODO: Comment this section out once the colleges below have been added to the database,
-        // TODO: so they are not added multiple times (prevent duplicate entries)
+        // Comment this section out once the colleges below have been added to the database,
+        // so they are not added multiple times (prevent duplicate entries)
+        /*
         db.addCollege(new College("UC Berkeley", 42082, 14068, 4.7, "ucb.png"));
         db.addCollege(new College("UC Irvine", 31551, 15026.47, 4.3, "uci.png"));
         db.addCollege(new College("UC Los Angeles", 43301, 25308, 4.5, "ucla.png"));
         db.addCollege(new College("UC San Diego", 33735, 20212, 4.4, "ucsd.png"));
         db.addCollege(new College("CSU Fullerton", 38948, 6437, 4.5, "csuf.png"));
         db.addCollege(new College("CSU Long Beach", 37430, 6452, 4.4, "csulb.png"));
+        */
 
         // Fill the collegesList with all Colleges from the database
         collegesList = db.getAllColleges();
@@ -66,7 +68,7 @@ public class CollegeListActivity extends AppCompatActivity {
         detailsIntent.putExtra("Name", selectedCollege.getName());
         detailsIntent.putExtra("Population", selectedCollege.getPopulation());
         detailsIntent.putExtra("Tuition", (float) selectedCollege.getTuition());
-        detailsIntent.putExtra("Rating", selectedCollege.getRating());
+        detailsIntent.putExtra("Rating", (float) selectedCollege.getRating());
         detailsIntent.putExtra("ImageName", selectedCollege.getImageName());
 
         startActivity(detailsIntent);
